enable use of geo_files and mapit
