+++ hurumap/templates/homepage.html
@@ -75,8 +75,12 @@
     <script src="{% static 'js/vendor/leaflet-0.6.4.js' %}"></script>
     <script src="{% static 'js/vendor/topojson.v1.min.js' %}"></script>
     <script src="{% static 'js/vendor/leaflet.label.js' %}"></script>
-    <script src="{% static 'js/map_mapit.js' %}"></script>
-    <script src="{% static 'js/profile_map_mapit.js' %}"></script>
+    <script src="{% static 'js/maps_static.js' %}"></script>
+    <script src="{% static 'js/profile_map.js' %}"></script>
+    {% if HURUMAP.use_mapit %}
+      <script src="{% static 'js/map_mapit.js' %}"></script>
+      <script src="{% static 'js/profile_map_mapit.js' %}"></script>
+    {% endif %}
     {% endblock %}
 
     <script>
