+++ hurumap/templates/profile/profile_detail.html
@@ -890,8 +890,10 @@
 {% block profile_javascript_libs %}
   <script src="{% static 'js/vendor/dom-to-image.min.js' %}"></script>
   {{ block.super }}
-  <script src="{% static 'js/map_mapit.js' %}"></script>
-  <script src="{% static 'js/profile_map_mapit.js' %}"></script>
+  {% if HURUMAP.use_mapit %}
+    <script src="{% static 'js/map_mapit.js' %}"></script>
+    <script src="{% static 'js/profile_map_mapit.js' %}"></script>
+  {% endif %}
 {% endblock profile_javascript_libs %}
 
 <script type="text/javascript">
